# Store replacement operations in order to do a single array resize