--- a/src/main/java/dev/blaauwendraad/masker/json/MaskingState.java
+++ b/src/main/java/dev/blaauwendraad/masker/json/MaskingState.java
@@ -50,16 +50,25 @@ public final class MaskingState {
         return message;
     }
 
+    /**
+     * Adds new delayed replacement operation to the list of operations to be applied to the message.
+     */
     public void addReplacementOperation(int startIndex, int endIndex, int maskLength, byte maskByte) {
         ReplacementOperation replacementOperation = new ReplacementOperation(startIndex, endIndex, maskLength, maskByte);
         replacementOperations.add(replacementOperation);
         replacementOperationsTotalDifference += replacementOperation.difference();
     }
 
+    /**
+     * Returns the list of replacement operations that need to be applied to the message.
+     */
     public List<ReplacementOperation> getReplacementOperations() {
         return replacementOperations;
     }
 
+    /**
+     * Returns the total difference between the masks and target values lengths of all replacement operations.
+     */
     public int getReplacementOperationsTotalDifference() {
         return replacementOperationsTotalDifference;
     }
@@ -72,8 +81,22 @@ public final class MaskingState {
                 " after: '" + new String(message, currentIndex, Math.min(10, message.length - currentIndex)) + "'";
     }
 
+    /**
+     * Represents a delayed replacement that requires resizing of the message byte array. In order to avoid resizing on
+     * every mask, we store the replacement operations in a list and apply them all at once at the end, thus making only
+     * a single resize operation.
+     *
+     * @param startIndex index from which to start replacing
+     * @param endIndex   index at which to stop replacing
+     * @param maskLength length of the mask to apply
+     * @param maskByte   byte to use for the mask
+     */
     public record ReplacementOperation(int startIndex, int endIndex, int maskLength, byte maskByte) {
 
+        /**
+         * The difference between the mask length and the length of the target value to replace.
+         * Used to calculate keep track of the offset during replacements.
+         */
         public int difference() {
             return maskLength - (endIndex - startIndex);
         }