--- a/src/main/java/dev/blaauwendraad/masker/json/util/FixedLengthTargetValueMaskUtil.java
+++ b/src/main/java/dev/blaauwendraad/masker/json/util/FixedLengthTargetValueMaskUtil.java
@@ -2,6 +2,8 @@ package dev.blaauwendraad.masker.json.util;
 
 import dev.blaauwendraad.masker.json.MaskingState;
 
+import java.util.Arrays;
+
 /**
  * Class containing utility methods to set a particular target value length to a fixed size mask. This can be used for
  * length obfuscation and ignoring escaping characters in string values.
@@ -15,47 +17,95 @@ public final class FixedLengthTargetValueMaskUtil {
      * Replaces a target value (byte slice) with a fixed length string consisting only of the mask bytes inside the
      * input bytes array.
      *
-     * @param fixedMaskLength   the length of the fixed-length mask byte string.
+     * @param maskLength        the length of the fixed-length mask byte string.
      * @param targetValueLength the length of the target value slice.
      * @param maskByte          the byte used for each byte in the mask
      */
     public static void replaceTargetValueWithFixedLengthMask(
             MaskingState maskingState,
-            int fixedMaskLength,
+            int maskLength,
             int targetValueLength,
             byte maskByte
     ) {
-        // Create new empty array with a length computed by the difference between requested fixed length and target
-        // value length.
-        byte[] newInputBytes = new byte[maskingState.getMessage().length + (fixedMaskLength - targetValueLength)];
+        int lengthDifference = maskLength - targetValueLength;
         int targetValueStartIndex = maskingState.currentIndex() - targetValueLength;
-        // Copy all bytes till the target value (including opening quotes of the string value to be replaced)
-        System.arraycopy(maskingState.getMessage(), 0, newInputBytes, 0, targetValueStartIndex);
-        // Start from beginning of target value and loop amount of required masked characters
-        for (int i = targetValueStartIndex; i < targetValueStartIndex + fixedMaskLength; i++) {
-            newInputBytes[i] = maskByte; // add masking characters
+        if (lengthDifference == 0) {
+            Arrays.fill(maskingState.getMessage(), targetValueStartIndex, maskingState.currentIndex(), maskByte);
+            return;
         }
-        // Append the rest of the original array starting from end of target value
-        System.arraycopy(
-                maskingState.getMessage(),
-                maskingState.currentIndex(),
-                newInputBytes,
-                targetValueStartIndex + fixedMaskLength,
-                maskingState.getMessage().length - maskingState.currentIndex()
-        );
-        maskingState.setMessage(newInputBytes);
+        maskingState.addReplacementOperation(targetValueStartIndex, maskingState.currentIndex(), maskLength, maskByte);
     }
 
     public static void replaceTargetValueWithFixedLengthAsteriskMask(
             MaskingState maskingState,
-            int fixedLength,
+            int maskLength,
             int targetValueLength
     ) {
         replaceTargetValueWithFixedLengthMask(
                 maskingState,
-                fixedLength,
+                maskLength,
                 targetValueLength,
                 AsciiCharacter.ASTERISK.getAsciiByteValue()
         );
+    }
+
+    /**
+     * Performs all replacement operations to the message array, must be called at the end of the replacements.
+     * <p>
+     * For every operation that required resizing of the original array, to avoid copying the array multiple times,
+     * those operations were stored in a list and can be performed in one go, thus resizing the array only once.
+     * <p>
+     * Replacement operation is only recorded if the length of the target value is different from the length of the mask,
+     * otherwise the replacement must have been done in-place.
+     *
+     * @param maskingState the current state of the {@link dev.blaauwendraad.masker.json.JsonMasker} instance.
+     */
+    public static void flushReplacementOperations(MaskingState maskingState) {
+        if (maskingState.getReplacementOperations().isEmpty()) {
+            return;
+        }
+
+        // Create new empty array with a length computed by the difference of all mismatches of lengths between the target values and the masks
+        // in some edge cases the length difference might be equal to 0, but since some indices mismatch (otherwise there would be no replacement operations)
+        // we still have to copy the array to keep track of data according to original indices
+        byte[] newMessage = new byte[maskingState.getMessage().length + maskingState.getReplacementOperationsTotalDifference()];
+
+        // Index of the original message array
+        int index = 0;
+        // Offset is the difference between the original and new array indices, we need it to calculate indices
+        // in the new message array using startIndex and endIndex, which are indices in the original array
+        int offset = 0;
+        for (MaskingState.ReplacementOperation replacementOperation : maskingState.getReplacementOperations()) {
+            // Copy everything from message up until replacement operation start index
+            System.arraycopy(
+                    maskingState.getMessage(),
+                    index,
+                    newMessage,
+                    index + offset,
+                    replacementOperation.startIndex() - index
+            );
+            // Insert the mask bytes
+            Arrays.fill(
+                    newMessage,
+                    replacementOperation.startIndex() + offset,
+                    replacementOperation.startIndex() + offset + replacementOperation.maskLength(),
+                    replacementOperation.maskByte()
+            );
+            // Adjust index and offset to continue copying from the end of the replacement operation
+            index = replacementOperation.endIndex();
+            offset += replacementOperation.difference();
+        }
+
+        // Copy the remainder of the original array
+        System.arraycopy(
+                maskingState.getMessage(),
+                index,
+                newMessage,
+                index + offset,
+                maskingState.getMessage().length - index
+        );
+
+        maskingState.setMessage(newMessage);
+        maskingState.setCurrentIndex(newMessage.length);
     }
 }