### Benchmark results (`./gradlew jmh -PjmhShort`):
Before:
```
Benchmark                                  (characters)  (jsonSize)  (maskedKeyProbability)  (obfuscationLength)   Mode  Cnt      Score     Error  Units
BaselineBenchmark.countBytes                    unicode       128kb                    0.01                  N/A  thrpt    4  34250.391 ▒ 420.174  ops/s
BaselineBenchmark.jacksonParseAndMask           unicode       128kb                    0.01                  N/A  thrpt    4    165.091 ▒   5.335  ops/s
BaselineBenchmark.regexReplace                  unicode       128kb                    0.01                  N/A  thrpt    4     51.704 ▒   3.354  ops/s
JsonMaskerBenchmark.jsonMaskerBytes    ascii (no quote)       128kb                    0.01                 none  thrpt    4   9915.660 ▒ 425.860  ops/s
JsonMaskerBenchmark.jsonMaskerBytes               ascii       128kb                    0.01                 none  thrpt    4   9305.348 ▒  62.295  ops/s
JsonMaskerBenchmark.jsonMaskerBytes             unicode       128kb                    0.01                 none  thrpt    4  12057.404 ▒ 419.343  ops/s
JsonMaskerBenchmark.jsonMaskerString   ascii (no quote)       128kb                    0.01                 none  thrpt    4   7925.503 ▒ 160.098  ops/s
JsonMaskerBenchmark.jsonMaskerString              ascii       128kb                    0.01                 none  thrpt    4   8053.798 ▒ 615.886  ops/s
JsonMaskerBenchmark.jsonMaskerString            unicode       128kb                    0.01                 none  thrpt    4   1987.428 ▒  17.674  ops/s
```

After:
```

Benchmark                                  (characters)  (jsonSize)  (maskedKeyProbability)  (obfuscationLength)   Mode  Cnt      Score     Error  Units
BaselineBenchmark.countBytes                    unicode       128kb                    0.01                  N/A  thrpt    4  34271.957 ▒ 337.662  ops/s
BaselineBenchmark.jacksonParseAndMask           unicode       128kb                    0.01                  N/A  thrpt    4    164.577 ▒   3.481  ops/s
BaselineBenchmark.regexReplace                  unicode       128kb                    0.01                  N/A  thrpt    4     65.975 ▒   0.297  ops/s
JsonMaskerBenchmark.jsonMaskerBytes    ascii (no quote)       128kb                    0.01                 none  thrpt    4   9992.484 ▒  62.882  ops/s
JsonMaskerBenchmark.jsonMaskerBytes               ascii       128kb                    0.01                 none  thrpt    4   8901.443 ▒  95.178  ops/s
JsonMaskerBenchmark.jsonMaskerBytes             unicode       128kb                    0.01                 none  thrpt    4  11286.123 ▒ 382.730  ops/s
JsonMaskerBenchmark.jsonMaskerString   ascii (no quote)       128kb                    0.01                 none  thrpt    4   8400.088 ▒ 109.190  ops/s
JsonMaskerBenchmark.jsonMaskerString              ascii       128kb                    0.01                 none  thrpt    4   8762.854 ▒  56.332  ops/s
JsonMaskerBenchmark.jsonMaskerString            unicode       128kb                    0.01                 none  thrpt    4   2408.000 ▒ 114.824  ops/s
```

### And for 2mb json with obfuscation enabled:
Before:
```
Benchmark                                  (characters)  (jsonSize)  (maskedKeyProbability)  (obfuscationLength)   Mode  Cnt     Score    Error  Units
BaselineBenchmark.countBytes                    unicode         2mb                    0.01                  N/A  thrpt    4  2272.636 ▒ 26.388  ops/s
BaselineBenchmark.jacksonParseAndMask           unicode         2mb                    0.01                  N/A  thrpt    4     7.887 ▒  0.106  ops/s
BaselineBenchmark.regexReplace                  unicode         2mb                    0.01                  N/A  thrpt    4     2.789 ▒  0.047  ops/s
JsonMaskerBenchmark.jsonMaskerBytes    ascii (no quote)         2mb                    0.01                    8  thrpt    4    14.366 ▒  0.069  ops/s
JsonMaskerBenchmark.jsonMaskerBytes               ascii         2mb                    0.01                    8  thrpt    4    13.617 ▒  0.086  ops/s
JsonMaskerBenchmark.jsonMaskerBytes             unicode         2mb                    0.01                    8  thrpt    4    15.337 ▒  0.780  ops/s
JsonMaskerBenchmark.jsonMaskerString   ascii (no quote)         2mb                    0.01                    8  thrpt    4    13.793 ▒  0.051  ops/s
JsonMaskerBenchmark.jsonMaskerString              ascii         2mb                    0.01                    8  thrpt    4    14.024 ▒  0.200  ops/s
JsonMaskerBenchmark.jsonMaskerString            unicode         2mb                    0.01                    8  thrpt    4    17.142 ▒  0.441  ops/s
```

After:
```
Benchmark                                  (characters)  (jsonSize)  (maskedKeyProbability)  (obfuscationLength)   Mode  Cnt     Score    Error  Units
BaselineBenchmark.countBytes                    unicode         2mb                    0.01                  N/A  thrpt    4  2271.825 ▒ 16.711  ops/s
BaselineBenchmark.jacksonParseAndMask           unicode         2mb                    0.01                  N/A  thrpt    4     7.997 ▒  0.239  ops/s
BaselineBenchmark.regexReplace                  unicode         2mb                    0.01                  N/A  thrpt    4     4.517 ▒  0.134  ops/s
JsonMaskerBenchmark.jsonMaskerBytes    ascii (no quote)         2mb                    0.01                    8  thrpt    4   341.738 ▒  4.737  ops/s
JsonMaskerBenchmark.jsonMaskerBytes               ascii         2mb                    0.01                    8  thrpt    4   301.066 ▒  7.211  ops/s
JsonMaskerBenchmark.jsonMaskerBytes             unicode         2mb                    0.01                    8  thrpt    4   366.295 ▒ 18.235  ops/s
JsonMaskerBenchmark.jsonMaskerString   ascii (no quote)         2mb                    0.01                    8  thrpt    4   285.796 ▒  6.253  ops/s
JsonMaskerBenchmark.jsonMaskerString              ascii         2mb                    0.01                    8  thrpt    4   278.421 ▒  5.353  ops/s
JsonMaskerBenchmark.jsonMaskerString            unicode         2mb                    0.01                    8  thrpt    4   145.581 ▒  5.877  ops/s
```